fix bug

+++ command_receiver.h
@@ -103,11 +103,11 @@ private:
     uint16_t sequence_;
     std::bitset<MAX_SEQ> packets_;
 
-    ssize_t total_delay_ = 0;
-    ssize_t avg_delay_ = 0;
-    ssize_t max_delay_ = 0;
-    ssize_t min_delay_ = 0;
-    ssize_t head_avg_delay_ = 0;;
+    int64_t total_delay_ = 0;
+    int64_t avg_delay_ = 0;
+    int64_t max_delay_ = 0;
+    int64_t min_delay_ = 0;
+    int64_t head_avg_delay_ = 0;;
     // var_delay_ = varn_delay_/n is variance
     uint64_t varn_delay_ = 0;
     // std_delay_ is standard deviation
